minor package refactor
